+++ framework/areg/trace/private/TraceMessage.cpp
@@ -119,7 +119,7 @@ void TraceMessage::logTrace( NETrace::eLogPriority logPrio, const char * format,
 inline void TraceMessage::_sendLog( unsigned int scopeId, NETrace::eLogPriority msgPrio, const char * format, va_list args )
 {
     LogMessage logData(NETrace::eLogMessageType::LogMessageText, scopeId, msgPrio, nullptr, 0);
-    logData.logMessageLen = String::formatStringList( logData.logMessage, NETrace::LOG_MESSAGE_IZE, format, args );
+    logData.logMessageLen = static_cast<uint32_t>(String::formatStringList( logData.logMessage, NETrace::LOG_MESSAGE_IZE, format, args ));
     TraceManager::logMessage( logData );
 }
